+++ Magic/CHANGELOG.md
@@ -37,6 +37,8 @@
    (Permission queries changed again, please let me know if anything weird happens, such as no permission to craft or use wands)
  - Allow using numbers for "slot" parameters in various places (such as the CheckInventory action)
  - Mobs spawned by the `monster` spell will attack for you
+ - The Damage action will now force-set the damager and damaged NMS references, so "magic" damage registers with
+   vanilla mob AI, possibly other plugins
 
 ## Fixes